+++ packages/experimental/Badge/src/CounterBadge/index.ts
@@ -1,0 +1,2 @@
+export * from './CounterBadge.types';
+export * from './CounterBadge';