+++ packages/experimental/Switch/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/switch
 
-This log was last generated on Thu, 04 Aug 2022 19:24:58 GMT and should not be manually modified.
+This log was last generated on Fri, 05 Aug 2022 23:22:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.1
+
+Fri, 05 Aug 2022 23:22:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.14.0
+- Bump @fluentui-react-native/interactive-hooks to v0.18.1
 
 ## 0.3.0